requirements.txt
django
djangorestframework
django-geojson
psycopg2-binary